Remove illegal characters from uploaded filenames
The characters  /, \, :, *, ?, ", <, >, and | are not allowed in
filenames. When a file was uploaded with these characters, it caused a
Aws::SESV2::Errors::BadRequestException error to be thrown when
attempting to send the email.

This commit removes these characters from the uploaded filename before
sending the email. This includes removing the characters from the
filename in the CSV file, if one is attached to the email and also for
submission to an S3 bucket.