Fix retry failed submission task
The task was assuming that there would be only one SolidQueue::Job
with the given active_record_id. However, if a job is retried, there
will be a SolidQueue::Job per retry. There will only be a
SolidQueue::FailedExecution for the Job for the the final retry.

Fix the rake task to get the last Job with the given active_record_id.