Access `page_slug` from page inside Step class
This removes the separate page_slug instance variable, which was
passed into the initializer for Step via the StepFactory. As the
original Page object is also passed in, we can reference it directly
from that object. This helps simplify the code and makes it easier for
future refactoring.

`page_slug` will always equal the `page.id` as the StepFactory matches
them, otherwise it raises an error.