Group Sentry events for submission bounced by the form ID
We tried adding the form ID to the Sentry event message in the hope of
Sentry creating separate issues per form ID. We did this so that we
can ignore the issues for certain forms where we know the bounce is
not of concern to us - usually due to the inbox having auto-replies
configured.

This didn't cause separate Sentry issues to be created because Sentry
will use the Stack trace rather than the formatted message when
grouping events. Giving the event a fingerprint that includes the
form ID should cause the events to be grouped by the form ID.

Including "{{ default }}" in the fingerprint means that Sentry will
use its default grouping rules first, but then further divide by the
form_id we pass in.