Add locale to scope
Instead of using a query string parameter, this commit adds the locale
to a route scope so that it will be included in the URL path.

It's set up to use the format form_slug.locale, so the resulting path
will look something like /1/apply-for-a-thing.en/. This aims for
consistency with the way GOV.UK handles these URLs on some pages
including detailed guides.