Improve validation of file answers in SES submission service
We recently got a "Number of files does not match number of completed
file questions" error from this service, and then had to spend a little
while trying to track down what exactly that error means and what
happened.

It turns out that this error can only really be raised if there are two
files which have the same name for their email attachments; so we should
rewrite the error message to say this.

In fact, the error we saw in production was where two files had the
original filename missing, so we also add a new error that uses the new
submission validation of the file. That will let us know if the original
filename is missing for any of the answered file questions in the form.