Add submission validation to Question::File
There are a couple of things we want to check about the answers for a
file question immediately before sending the submission; however we
don't keep the file data in the answer store, so we can't use the same
validations at this point.

This commit changes the existing validations so they don't run for the
submission context, and adds new validations that only run for the
submission context.

This is mainly useful for the scenario where somehow we've stored the
uploaded_file_key but not the original_filename.