Default runner to talk to forms-admin rather than forms-api
We have switched our deployed versions of forms-runner to make
requests to the forms-admin API endpoints through the internal load
balancer rather than to forms-api by setting this environment variable
for the ECS tasks.

Default the local configuration to make requests to forms-admin rather
than forms-api to match the deployed configuration.

Because we need to use the hostname for the internal load balancer in
the deployed environments, we need to keep a separate setting for this
URL to the `forms_admin.base_url` setting which is used to construct
links to the admin interface. We may want to consider renaming this
setting in future to make it not refer to forms-api.