Use keyword arguments in path helpers
We were using the form_submit_answers_path and form_page_path helpers in
our tests with positional arguments, which made it hard to tell what
each argument did.

One consequence of this was that each use of the
form_submit_answers_path helper in our tests had a `1` as one of its
arguments, which didn't seem to correspond to any of the possible params
in `config/routes.rb`. This caused a test failure with the new default
locale param - the helper was reading this `1` as the locale value.

This PR:

- removes the redundant `1` parameter from form_submit_answers_path
- uses keyword arguments for all uses of the helpers, so we're explicit
about what we're passing.