Delete all submissions after 30 days
We already delete submissions 7 days after delivery was attempted if
we have not received a notification that the email bounced.

We have a retention policy that we will only keep submission data for
a maximum of 30 days, so we need to delete submissions after 30 days
regardless of whether the email was delivered or not.