Refactor CsvGenerator to not write to a file
When we first started attaching CSV files to submission emails, we
were using Notify. The Notify client needed a File object to upload
the file for the Notify email to link to.

Now that we send emails using SES, we don't need to write to a
temporary file, and can instead just pass the contents for the file
and the filename to the mailer to set as an attachment on the email.

Similarly for S3 submissions, we just need to provide the file content
to the S3 client, so we don't need to write to a temporary file.