Update spec to test order of submission attachments
We want to ensure that we don't change the order of the submission
email attachments when both CSV and JSON files are attached. This is
because some automations may rely on the order of attachments. If the
order wasn't guaranteed, consumers should be able to handle this but
we can avoid potential issues by ensuring a consistent order.