Support add another answer in submission JSON
Support questions where the user can answer multiple times (add
another answer) when generating the submission JSON. We want to handle
this by including the question multiple times in the answers array of
the JSON, once for each answer provided by the user.

This involves moving where we generate the items in the answers array
of the JSON to the Step class, and adding an implementation for the
RepeatableStep subclass.