Send submission as JSON if submission_type is s3_with_json
We've added "s3_with_json" as a possible submission_type for a form.
If the form has this submission type, send the file to s3 as a JSON
file.

If the submission type is "s3", send the file as a CSV.

We will do some work later to separate the submission type (s3/email)
from the submission format (csv/json).