Refactor Form to work with v2 form documents
We want to bring the models in forms-runner closer to the form document
representation used in the API wire format. This change tweaks the Form
model to work with a v2 form document, getting us a little closer to
parity.

Some conversion is still needed to get pages instead of steps, but it is
done by the Form model now, and we can continue improving things
incrementally in future commits.