Update page ID regex to allow external IDs
This is to allow us to switch pages/steps to use external IDs rather
than the database ID.

Move the constant from `step_factory` directly into into the route
constraint. Remove the tests from `step_factory_spec` as the regex is
not applied in the same way for routes as these tests assume, as the
routing code automatically applies anchors to the regex. The
application of the page_slug constraint is tested adequately in
`page_controller_spec`.

See:
https://guides.rubyonrails.org/routing.html#segment-constraints

Added a separete regex for use when matching a page_id for logging
which does use anchors to match the entire string.