Use Rails cache store for session store in development and test
Storing form documents in the store busts the maximum cookie size, so we
can't use that for the session store.

In production and other AWS environments, we use Redis, and this commit
shouldn't change that.

For local development, usually devs should have Redis running, but in
case they don't we change the fallback from the Rails default of the
cookie store to a configuration which uses the Rails cache, which is
configured to be an in-process memory store. One annoying side-effect of
this is that sessions will now be lost whenever the server is restarted;
so this commit also adds a warning to devs about that. But hopefully
most devs are using Redis so this shouldn't be a problem.

For tests we need something that persists more than the local cache
strategy for feature and system specs, otherwise the answers get lost
between page navigation events, so we also change the test environment
configuration to use the memory store. To try and prevent any nasty side
effects from this we also add a before hook to all specs to clear the
cache before the spec starts.