wip! Cache form documents for form filler sessions
If a form is archived, the current behaviour is that the form filler
will immediately see a page explaining that the form has been archived.
This commit preserves this behaviour; if the form is archived and the
form filler is filling out the live form they will not be able to
continue their session even though the form document is present in their
session.

If a form is submitted, the form document is removed, so that a new
session will use the latest version of the form (if any updates are
present).

Currently this does result in a lot of extra usage of Redis; we could
look at deduplication in future.