Allow using database ID in URLs after FormDocument migration
We are going to migrate FormDocuments to use Page external IDs for
the steps instead of the database IDs.

To avoid breaking existing user sessions we're also temporarily adding
in a database_id attribute. If someone attempts to POST to a page URL
that contains a database ID just after the migration has been run, we
can use the database_id attribute to find the correct step.

This means that users shouldn't see errors when they try to submit
an answer just after the migration has run.