Retreive answers from store using external ID or database ID
To prepare for switching step IDs to be external IDs, we need to
ensure that we are able to retreive existing answers from user's
sessions that were stored using the database ID as a key.

To acheieve this, we will temporarily be adding a `database_id`
attribute to form documents when we switch the id to be the external
ID. When we switch the IDs, we will be storing new answers using the
external ID but will fallback on getting existing answers using the
`database_id` if no answer is returned.