Convert selection questions to use English value
Change selection questions to use value as the selection attribute and
modify show to use name based on the value.

This is the main change to the codebase to support routing in Welsh
forms.

Selection questions now store the English text in the value attribute
and the Welsh text in the name attribute.

Because routing is based on the English value all forms of routing will
continue to work in both languages.

We change how the selection question is displayed in the check your
answers page to use the Welsh value, if the locale is Welsh.

Submissions are generated using the English value. This spike hasn't
looked at submiissions but we'll need to make an extra request for the
current form in english when generating a submission, if we want the
submission to use the English questions text.