Fix typo in Welsh translation
The original translation for this field read 'The answer must be greater than 10 or equal to 0' , where the '10' was a typo. We've clarified this with the translators and the correct translation reads as 'The answer must be greater than 0 or equal to 0', which is the clearest way to say greater than or equal to 0 in Welsh.