Add params to FormDocumentResource
When making get requests to the Admin API, we want to be able to pass
params to the request to set the language.

This commit passes params in #get and #find to the parent class.

get uses the name options, to match the parent method.

It's not clear where we are using find but we update it anyway.

To keep things neat, we also change find's path to use the
collection_name method.