Add comments to point to updating the WAF rules
We have a WAF rule that allows large request bodies for the submit
page URL that uses a regex to match on the URL. Add comments on the
constraints we apply to the page route to remind to update the WAF
rules if we change the allowed values.

Add regex constraints on the form_id  and form_slug that matches the
regexes in the WAF rule with a comment to remind to update the WAF
rule if this changes.