Add language helpers to Form
FormDocuments can have a language attribute, which is used to determine
whether the form is in English or Welsh.

This attribute doesn't exist in all form documents so we need to account
for that.

This commit adds helpers to Form to determine whether the form is in
English or Welsh.