Ensure that action shasums match the release in the comment
I noticed that in PR #1768 Dependabot was not updating the version
number in the comment for the checkout action when updating the SHA, I
suspect this is because the current SHA is not the same as for the
v5.0.0 release.

This commit changes the actions/checkout SHA to v5.0.0 release commit,
hopefully this will let Dependabot sort itself out.