Add method that returns whether autocomplete is shown
If users will be asked to provide another answer if 'None of the
above' is selected, we need to know whether an autocomplete will be
shown to determine whether the question is on the same page in a
conditional reveal or on a new page. Add this method to re-use this
code.