Add method to get whether there is a none of the above question
Add a method to the selection question model to determine if there is
an additional question that should be shown to the user when they
select 'None of the above'.

Use `Struct` rather than `OpenStruct` in the tests to better mock out
ActiveResouce::Base which throws a NoMethodError if we try to access
an attribute that doesn't exist.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>