Add validation for none of the above question
Add validation for the none of the above question when it is displayed
for checkboxes or radios. When an autocomplete component is displayed,
the field will be on a separate page - validation for this will be
handled in a subsequent commit.

The error messages are the same as those for a text answer, so the
Welsh translations are directly copied from those error messages.