Clear none_of_the_above_answer if None of the above not selected
Add a before_validation callback to clear the
`none_of_the_above_answer` for a selection question if 'None of the
above' is not the selected option. This is because we don't want to
include it in the saved answer in this case. This also means that we
no longer need to check the selected option
when applying length validation for `none_of_the_above_answer`.