Use selection factory in selection_spec
We've defined the answer_settings in the factory to use Struct rather
than OpenStruct for the answer_settings to better mock the behaviour
of ActiveResouce::Base which raises a NoMethodError when trying to
access attributes that don't exist.

Use the factory so we don't need to define this behaviour in multiple
places.