Add component for check your answers table
We need to make some changes to display the answer for the additional
question for a selection question that asks another question when 'None
of the above' is selected in the check your answers table.

Move the code that renders the summary list into a component so it's
easier to test.

Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>