Override answered? method for selection questions
For a selection question with more than 30 options where we display an
autocomplete component, we'll want to show the extra question if
'None of the above' is selected on a new page. In the case we show
a question when none of the above is selected, we don't want users to
be able to go to the next question in the form until it is answered.

Override the `answered?` method to return false if there is a 'None of
the above' question configured and it hasn't been answered.