Refactor helper method to get hidden preview mode
Rather than passing in the question text suffix from the controller to
the ViewComponent, pass across the Mode so that the ViewComponent can
call the helper method directly to get the required hidden text span
to append to the question name.

This is needed as we're adding a separate page for the None of the
question for long list questions which will use a different field
rather than the `question_text` as the page title, and we do not need
to use a ViewComponent in the view for this so its preferable to
directly call the helper method from application_helper rather than
having the logic split between application_helper and QuestionBase.