Validate the none_of_the_above answer for autocomplete questions
When an autocomplete component is shown for a selection question when
it has more than 30 options, we show the question for when
"None of the above" is selected on a different page. To account for
this, we do not validate the presence of the none_of_the_above_answer
on the question page if the question is an autocomplete question.

We use the same question model for submitting the answer on the
separate page for the "None of the above" question, so we need to
pass in a context to the validation to make sure the presence of the
none_of_the_above_answer is validated when the question is mandatory.