Skip validating the none_of_the_above_answer for autocomplete
When an autocomplete component is shown for a selection question when
it has more than 30 options, we show the question for when
"None of the above" is selected on a different page. To account for
this, we do not validate the presence of the none_of_the_above_answer
on the selection page if the question is an autocomplete question.

Allow passing in a validation_context to the validation method that
will skip the validation of the none_of_the_above_answer.

We want to validate the none_of_the_above_answer in the default
context in case we have a future use for loading in questions and
validating, rather than just the current case where we validate the
model when a question is answered.