Add route for none of the above question page
Add a separate /none-of-the-above route for asking the question when
"None of the above" is selected for a selection question that uses the
autocomplete component. Use the same `Question::Selection` model for
the form on this page. We need to set the `selection` for the model to
have "None of the above" selected as this won't be loaded
automatically from the answer_store as `answered?` returns false for
the question. We know what the selection is though, so rather than
change the step loading logic, we can just pre-set.

When the step is saved, pass in the context `none_of_the_above_page`
so that the none_of_the_above answer is validated correctly.