Redirect to the none of the above question page
When the page for a selection question is submitted, and it is an
autocomplete component with "None of the above" selected, redirect to
the dedicated "None of the above" question page.

Do not call LogEventService#log_page_save in this case as we don't
want 2 save events for the same question, and we also only want to
send the "Started" metric to CloudWatch once for this form session if
this is the first question in the form.

Pass in the `skip_none_of_the_above_question_validation` validation
context when saving the step so that we don't validate the presence of
a value for the none of the above answer on the selection page.