Show back link when there are validation errors
We weren't showing the back link when rendering validation errors in
response to the POST request to submit a question. Move setting the
back_link to the common method for setting up the instance variables
and make it more uniform with how we set the other instance variables.

Also improve the readability of the method while we're here.